fix: tier 0-2 low-severity bug sweep — cache, lifecycle, storage, http, clock threading (#68)
* fix(cache): a refusing tier never fails the business read

- createCacheStack routes every get/set/del through bestEffort(): a tier
  that throws is a tier that did not answer, so an over-budget LRU entry
  (X_CACHE_TOO_LARGE) or a dead Redis costs the entry, never the read
- load() stays unguarded — it is the business read
- absorbed refusals land in recentTierFailures() (last 100, newest first,
  tier/op/key/code) plus one cache.tier.failed warn; resetTiers() clears it
- 8 stack tests (7 red without the guard) + 12 log tests; README, CLAUDE.md,
  wiki and CHANGELOG in lockstep


* fix(http,storage): a mis-encoded path is a 400, not a 500

- router.ts calls decodeURIComponent only through a guarded decodeSegment;
  a segment that will not decode answers the new MatchResult variant
  { reason: 'path-invalid', segment } instead of throwing a bare URIError
- pipeline turns that into X_PATH_INVALID -> 400 (title, ERROR_STATUS row,
  pathInvalid() factory, wiki row, manifest); it used to reach factsOf as
  X_INTERNAL, answer 500 and page the on-call for a client typo
- only the branch that would have decoded fails: a path reaching no param
  or wildcard is the 404 it always was, precedence unchanged
- signed-url.ts parseConstraints decodes through the same guard, so a %ZZ
  key is 'malformed' and verifySignedUrl keeps its never-throws promise
- 15 new tests, 9 red without the fix; CHANGELOG, both CLAUDE.md in lockstep


* fix(http): handle() always answers a Response

- guard the finalize loop and the recover stage in finalize.ts: a stage
  that refuses the response it was handed degrades to the new
  X_PIPELINE_FINALIZE_FAILED (500) instead of rejecting handle()
- the chain runs a second pass over the problem document, so the
  degraded answer still carries x-request-id, CORS and security headers;
  a second failure keeps its 500 and stops
- the failure travels through the recover stage, so reporting, logging
  and the overlay each keep one call site


* fix(seo): one bad date no longer takes the whole feed down

- buildFeed parsed item timestamps straight into new Date(...).toISOString(),
  so an unparseable published threw a bare RangeError out of the feed route
- Math.max(...times) spread one argument per item, so a feed crashed in
  proportion to how well the blog did; the empty branch called Date.now()
- feed-dates.ts is now the only place a feed timestamp is parsed or formatted:
  an unusable date is absent (never Invalid Date, never invented), it sorts
  last instead of returning NaN from the comparator, and "now" is a Clock seam
  via buildFeed(channel, items, { clock })


* fix(http,auth,i18n): bound the IP-keyed limiters, own-property vars

- both in-memory rate limiters carry `forgetAtMs` and sweep entries that
  answer exactly as a missing one; a cap backstops what the sweep can't
  claim, evicting nearest-to-forgotten first (auth ranks a live lockout
  last) so filling the table is not a rate-limit bypass
- `memoryRateLimitStore({ maxKeys })` / `policy.maxKeys` + an observable
  `size` on both, so the bound is asserted, not claimed
- `interpolate` reads only own properties: `{constructor}` rendered a
  function's source into the page; fast path now tests both braces so
  `}}` un-escapes with no `{` in sight


* fix(core): dedupe timingSafeEqual, register schema's error codes

- Move the byte-identical timingSafeEqual out of @ultimat3/auth and
  @ultimat3/storage into @ultimat3/core (tier 0, both packages already
  depend on it); each package re-exports it so no import changes.
- @ultimat3/schema's X_VALIDATION_FAILED/X_SCHEMA_UNSUPPORTED titles were
  registered only by the CLI, so a non-CLI process rendered the humanised
  fallback. core now registers a tested duplicate at import time
  (schema-error-codes.ts), pinned equal to schema's own declarations by
  schema-error-codes-pin.test.ts in @ultimat3/cli, the lowest tier that can
  import both. The CLI's now-redundant registration is removed.


* fix(money,schema,entity): one Money declaration, minor is a number

Money was three structural restatements of one shape, and entity's copy
had a `bigint` minor. That was not stylistic: JSON.stringify throws on a
bigint, so returning a decoded row from an action crashed the response,
t.money — the node that becomes the OpenAPI contract — rejected the row
this framework's own driver produced, and realtime normalised the same
column to a number, so two readers of one column disagreed.

- @ultimat3/schema (tier 0) owns the single declaration; money's `Money`
  and entity's `MoneyValue` are aliases of it, not copies of its shape.
- Both fields are `readonly`, which the type's own docs already asked for.
- The column stays `bigint`: a stored minor unit past ±2^53 is refused
  where it is read, naming the value — never rounded, never carried as a
  bigint that crashes three layers later. Same refusal realtime makes.
- Additive on the write side: `MoneyInput` still takes a bigint, and
  `narrowMoney` is called by BOTH drivers so a row's money never depends
  on which one produced it — memoryRepo previously stored the caller's
  bigint verbatim, the one row JSON.stringify refused.
- type-pins.ts fails the build on a re-declared alias, a widened `minor`,
  or a dropped `readonly`.


* fix: low-severity cleanup — cache stats, drain leak, storage round-trip, clock threading

- LruCache.clear() now resets hits/misses/evictions along with entries/bytes
- core lifecycle: a timed-out drain waiter no longer lingers in idleWaiters;
  adds test-only idleWaiterCount()
- storage: StorageObject carries cacheControl/metadata so get()/list() round-trip
  what put() wrote; local driver's sidecar parser no longer drops them; s3 driver's
  list() content-type gap is now a documented API limitation, not silent
- http: HttpServer.stop() unregisters its shutdown hooks in a finally, so a
  throwing drain() can't leave them registered
- entity + db: nine ambient new Date() "now" reads replaced with
  @ultimat3/core's systemClock.now(), closing the "only clock.ts calls new
  Date()" gap (behavior-preserving; now?: Date overrides unchanged)


---------

+++ packages/entity/src/cursor.ts
@@ -12,7 +12,14 @@ import { invariantViolated } from './errors';
 import type { QueryPlan } from './tenancy';
 import type { AnyColumn, ColumnKind } from './types';
 
-const MONEY_PARTS: Readonly<Record<string, ColumnKind>> = { minor: 'bigint', currency: 'char' };
+/**
+ * The kind a money part is *revived* as, which is the kind the row property holds — not the
+ * physical column's. `<p>_minor` is a `bigint` column, but `MoneyValue.minor` is a `number`
+ * (`@ultimat3/schema` owns that declaration), and a cursor whose value came back a `bigint`
+ * would compare against a `number` property in the memory driver and mint a seek bind of the
+ * wrong type in the other. The narrowing itself is guarded once, where the row is decoded.
+ */
+const MONEY_PARTS: Readonly<Record<string, ColumnKind>> = { minor: 'integer', currency: 'char' };
 
 /** Resolves `price.minor` as well as `title`; money is the one property with two parts. */
 const partsOf = (path: string): { readonly property: string; readonly part?: string } => {
